Added sprockets imports for jquery and foundation.
Initially jquery and foundation were just "baked in" to this project, which meant I couldn't update either easily. So I brought them in with bower, and added sprockets imports to pull in the copies from ```bower_components``` instead of the ```vendor``` copies.

This required modifying all the ```require``` statements in ```app.js.coffee``` to refer to the new locations. I also found that I needed to provide "full" paths in order for the requires to work. That doesn't seem entirely right to me, but it was what worked.