Sketched out several versions of the image gallery.
This commit captures several sketches of a possible image gallery. When I ran it past Sue, she liked the Flexbox version, which I think will be the easiest to work with, so I'll delete the others. I wanted to capture all this in git before I started to pull stuff out, though.