Explicitly specified columns, and added a "by line".
I explicitly specified the column setup using the Foundation grid tools. I feel like maybe I shouldn't or didn't need to do that, but it did seem to improve the look and feel, so I'm not sure.

I also added the "by" entry to each of the titles in the HAML since adding it in the JS didn't seem to work, and I think I like it better here anyway as it keeps the content together in one place.