Commented out all the drop downs.
For the moment we don't have the separate galleries, and the only "About" page we're keeping right now is the CV page, so I commented all the rest out. I also removed the right justification for the CV link in the top bar since that left that link really isolated over on the right, especially on a wide screen.