Converted raw HTML (in a .markdown file) to HAML.
I had raw HTML in a file that purported to be Markdown but wasn't. I converted all that to HAML and changed the file extension. What I should probably do next is move all the content into head matter, and then use looking in HAML to build out all the formatting and structure.